feat: use splash screen to load font instead
